% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_leroux.R
\name{load_leroux}
\alias{load_leroux}
\title{Load Leroux distribution}
\usage{
load_leroux()
}
\value{
No return value.
}
\description{
Register the Leroux CAR distribution so that it can be used in NIMBLE model
code.
}
\keyword{internal}
